feat: add camera utilities and update README instructions for webcam usage
