+++ FlowerAI_KeyPoints/server.py
@@ -4,7 +4,7 @@
 # Servidor Flower (plano de control) para KeyPoints (Keypoint R-CNN ResNet50-FPN)
 # Publica:
 #  - conf_thr (float, default 0.5)
-#  - input_size (int, default 320)
+#  - input_size (int, default 640)
 #  - max_frames (int, opcional; 0 => sin límite)
 #  - draw (int bool 0/1)
 
@@ -17,9 +17,9 @@ import flwr as fl
 def build_argparser() -> argparse.ArgumentParser:
     p = argparse.ArgumentParser(description="Servidor Flower - KeyPoints (Keypoint R-CNN ResNet50-FPN)")
     p.add_argument("--address", type=str, default="0.0.0.0:8080", help="Dirección host:puerto del servidor Flower")
-    p.add_argument("--rounds", type=int, default=1_000_000_000, help="Número de rondas (mantener proceso activo)")
+    p.add_argument("--rounds", type=int, default=1, help="Número de rondas")
     p.add_argument("--conf", type=float, default=0.5, help="Umbral de confianza")
-    p.add_argument("--input-size", type=int, default=320, help="Lado corto de la imagen de entrada")
+    p.add_argument("--input-size", type=int, default=640, help="Lado corto de la imagen de entrada")
     p.add_argument("--max-frames", type=int, default=0, help="Límite de frames a procesar (0 = sin límite)")
     p.add_argument("--draw", type=int, choices=[0, 1], default=0, help="Dibujar skeleton en el cliente (0/1)")
     return p
